Dynamically import the correct contact state
